update_dim_tables_cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.30.0
google-cloud-storage==3.1.0